refactor(utils): split questionnaire-api-sync under the line cap (#544)
Extraction only, zero behavior change. Mirrors the established
poll-api-sync-* sibling pattern: the per-question/section CRUD helpers
move to questionnaire-api-sync-questions.ts and the conditional-content
helpers to questionnaire-api-sync-conditionals.ts (the two are mutually
recursive, resolved via hoisted function declarations). The original
module keeps saveQuestionnaireIncremental and re-exports every prior
symbol so importers and tests need no changes.
